requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
contourpy==1.3.3
coverage==7.10.5
cycler==0.12.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
fastjsonschema==2.21.2
Flask==3.1.1
fonttools==4.59.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.2
openai==1.107.2
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
postgrest==1.1.1
priority==2.0.0
pyarrow==22.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
Quart==0.20.0
quart-cors==0.8.0
realtime==2.7.0
referencing==0.37.0
rpds-py==0.28.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
